chore: move search topbar to dedicated component
