+++ src/views/search/search-page.module.scss
@@ -20,124 +20,3 @@
 	max-width: var(--max-width_l);
 	padding: var(--site-spacing);
 }
-
-.topBar {
-	display: grid;
-	gap: var(--spc-4x);
-
-	@include from($tabletSmall) {
-		grid-template-rows: repeat(3, auto);
-		grid-template-columns: repeat(2, auto);
-	}
-
-	@include from($tabletLarge) {
-		gap: var(--spc-8x);
-		grid-template-rows: 1fr;
-		grid-template-columns: 1fr repeat(2, auto);
-	}
-}
-
-.topBarDivider {
-	height: 1px;
-	width: 100%;
-	background: var(--background_disabled);
-
-	@include from($tabletSmall) {
-		grid-column: 1 / 3;
-	}
-
-	@include from($tabletLarge) {
-		grid-column: unset;
-		height: 100%;
-		width: 1px;
-	}
-}
-
-.topBarButtonsContentToDisplay {
-	display: flex;
-	flex-direction: row;
-	justify-content: center;
-	align-items: center;
-	gap: var(--spc-2x);
-
-	@include from($tabletSmall) {
-		grid-column: 1;
-		justify-self: start;
-	}
-
-	@include from($tabletLarge) {
-		grid-row: unset;
-		grid-column: unset;
-	}
-}
-
-.topBarSmallTabletButtons {
-	display: flex;
-	flex-direction: row;
-	justify-content: center;
-	align-items: center;
-	gap: var(--spc-2x);
-
-	// Setup for the small tablet buttons
-	grid-column: 2;
-	justify-self: end;
-
-	@include from($mobile) {
-		display: none;
-	}
-
-	@include from($tabletSmall) {
-		display: flex;
-	}
-
-	@include from($tabletLarge) {
-		display: none;
-	}
-}
-
-.topBarSmallTabletButtonsToggle {
-	display: flex;
-	flex-direction: row;
-	justify-content: center;
-	align-items: center;
-	gap: var(--spc-2x);
-}
-
-.filterIconContainer,
-.filterIconContainer svg {
-	height: 100%;
-	width: 100%;
-}
-
-.searchbarRow {
-	display: flex;
-	flex-direction: row;
-	width: 100%;
-	gap: var(--spc-2x);
-	justify-content: center;
-	align-items: center;
-
-	@include from($tabletSmall) {
-		grid-column: 1 / 3;
-	}
-
-	@include from($tabletLarge) {
-		grid-column: unset;
-	}
-}
-
-.searchbar {
-	width: 100%;
-}
-
-.searchButton {
-	height: calc(
-		var(--p_medium_line-height) +
-			calc(var(--search-form-field_padding-vertical) * 2)
-	);
-	width: calc(
-		var(--p_medium_line-height) +
-			calc(var(--search-form-field_padding-vertical) * 2)
-	);
-	flex-shrink: 0;
-}
